fix: 직렬화 타입 지정을 위해 지역정보 dto infra 모듈로 변경
